Fix tximeta cache issue
tximeta attempts to use information for the wrong linked txome confusing e.g. mouse and human annotations.

To fix this we now set the BiocFileCache location to be "tximeta.dir/cache" when running pipeline_salmon.py. This ensures use of the correct linked transcriptome when running SummarizeToGene as there will only be one record in the cache.